Only need convert counter to LONG
Reduce conversion necessary for optimization.

+++ src/CxbxKrnl/EmuDSoundInline.hpp
@@ -1513,7 +1513,7 @@ inline HRESULT HybridDirectSoundBuffer_SetMixBinVolumes_8(
                 }
             }
             if (counter > 0) {
-                Xb_volumeMixBin = LONG((float)volume / (float)counter);
+                Xb_volumeMixBin = volume / (LONG)counter;
                 hRet = HybridDirectSoundBuffer_SetVolume(pDSBuffer, Xb_volume, EmuFlags, nullptr,
                                                          Xb_volumeMixBin, Xb_dwHeadroom);
             } else {
